refactor_estrutura
