+++ funcionalidades/espaco_de_ofertas.py
