+++ funcionalidades/ranking.py
